Introduced string extentions for single line operations.
Finally decided to introduce extending the ruby String class for
single line operations. This make the code tighter in a lot
of places and removes the need to create and invoke a generator
for those simpler operations (make the simple things simple and
the difficult things possible). The readme generator has been
modified to work with the new string extensions, but I need to
add more content to discuss how the division of responsibilities
falls between the String extensions and the Generator.