Adding more methods to the String extension, bullet, number, code and code_span.
Adding the code span has now introduced a dependency on the
MarkMaker module constants. I moved the mark_maker_string.rb file
up a level to try and better reflect the actual dependencies and
not imply that the string was actually part of the MarkMaker
module itself. I'm still not sure that I'm handling this right.
Maybe there needs to be an additional more neutral location for
the constants so that you don't need the MarkMaker module in order
to use the String extensions on their own?