Added headers 4-6. Fixed bug in image generation.
Added methods header4, header5, header6. Would like this to be
consolidated into something more meta programming in the
future but wanted to at least get to parity with the Markdown
spec and put some tests in place. Also noticed a glitch
where the image title was malformed and fixed that issue.